src/nptv6: relocate inline tests to nptv6_tests.rs
Same playbook. 349-LOC inline test block extracted to sibling nptv6_tests.rs.

After: nptv6.rs drops from 582 → 236 LOC.

cargo build --release: clean. cargo test --release: 878 passed.
